LPS-57368 Remove TransactionHandlerUtil which was added for arquillian extension support (no longer needed), to avoid having some confusing api of programmatic transaction control.
