+++ portal-impl/src/META-INF/base-spring.xml
@@ -81,9 +81,13 @@
 		<constructor-arg ref="liferayTransactionManager" />
 		<constructor-arg value="false" />
 	</bean>
-	<bean class="com.liferay.portal.spring.transaction.TransactionHandlerUtil">
-		<property name="platformTransactionManager" ref="liferayTransactionManager" />
-		<property name="transactionExecutor" ref="transactionExecutor" />
+	<bean id="com.liferay.portal.kernel.transaction.TransactionInvokerUtil" class="com.liferay.portal.kernel.transaction.TransactionInvokerUtil">
+		<property name="transactionInvoker">
+			<bean class="com.liferay.portal.spring.transaction.TransactionInvokerImpl">
+				<property name="platformTransactionManager" ref="liferayTransactionManager" />
+				<property name="transactionExecutor" ref="transactionExecutor" />
+			</bean>
+		</property>
 	</bean>
 	<bean class="com.liferay.portal.transaction.TransactionLifecycleManagerInitializer">
 		<property name="transactionLifecycleListeners">
